add useProposalsTab custom hooks
